Some changes to the PPU to increase efficiency. During VBLANK PPU jumps ahead as many ticks as it can all in one go rather than advancing slowly in a loop. I also changed the way the front end draws the frames, which vastly improved performance

+++ src/Emulator/ppu.h
@@ -35,7 +35,7 @@ class PPU
     boost::chrono::high_resolution_clock::time_point intervalStart;
 	std::atomic<bool> limitTo60FPS;
 
-    uint32_t ppuClock;
+    uint64_t ppuClock;
     int16_t dot;
     int16_t line;
     bool even;
